+++ src/query_2.c
@@ -35,7 +35,7 @@ LONG_list top_most_active_aux(GHashTable* com_user, int N) {
 
 	for (i=0;i<size;i++){
 		user = g_array_index(users,User,i);
-		set_list(res,i,(long)getUserID(user));
+		set_list(res,i,getUserID(user));
 
 	}
